Controllers and delegates refactor
