+++ Movies/Movies/DataSources+Delegates/MoviesCollectionViewDelegate.swift
@@ -8,16 +8,29 @@
 
 import UIKit
 
-class MoviesCollectionViewDelegate: NSObject, UICollectionViewDelegateFlowLayout {
-  var cellWidth: CGFloat
-  var newPosterHeight: CGFloat
-  var margin: CGFloat = 16
-  var columns: CGFloat = 2
-  var posterHeight: CGFloat = 240
-  var posterWidth: CGFloat = 160
-  var movieInfoContainerHeight: CGFloat = 66
-  var updateDelegate: MoviesCollectionViewUpdateDelegate
-  var itemSelectDelegate: CollectionViewdidSelectItemDelegate
+final class MoviesCollectionViewDelegate: NSObject, UICollectionViewDelegateFlowLayout {
+  
+  // MARK: Properties
+  
+  let cellWidth: CGFloat
+  
+  let newPosterHeight: CGFloat
+  
+  let margin: CGFloat = 16
+  
+  let columns: CGFloat = 2
+  
+  let posterHeight: CGFloat = 240
+  
+  let posterWidth: CGFloat = 160
+  
+  let movieInfoContainerHeight: CGFloat = 66
+  
+  let updateDelegate: MoviesCollectionViewUpdateDelegate
+  
+  let itemSelectDelegate: CollectionViewdidSelectItemDelegate
+  
+  // MARK: Initialization
   
   init(frameWidth: CGFloat, updateDelegate: MoviesCollectionViewUpdateDelegate, itemSelectDelegate: CollectionViewdidSelectItemDelegate) {
     self.cellWidth = (frameWidth - (margin * (columns + 1))) / columns
@@ -26,22 +39,23 @@ class MoviesCollectionViewDelegate: NSObject, UICollectionViewDelegateFlowLayout
     newPosterHeight = posterHeight * cellWidth / posterWidth
   }
   
+  // MARK: Table view delegate
+  
   func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
     return CGSize(width: cellWidth, height: newPosterHeight + movieInfoContainerHeight)
   }
   
   func collectionView(_ collectionView: UICollectionView, willDisplaySupplementaryView view: UICollectionReusableView, forElementKind elementKind: String, at indexPath: IndexPath) {
-    
     if updateDelegate.canShowFooter() {
       view.isHidden = false
       updateDelegate.loadMoreMovies()
     }
-    
   }
 
   func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
     itemSelectDelegate.didSelectIndexPath(indexPath)
   }
+  
 }
 
 protocol CollectionViewdidSelectItemDelegate: class {
